music/serializers.py
from rest_framework import serializers
from .models import Song
from django.contrib.auth.models import User

class SongSerializer(serializers.ModelSerializer):
    class Meta:
        model = Song
        fields = ['id', 'title', 'artist', 'youtube_url', 'thumbnail', 'video_id', 'file_path']
class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'username']

