requirements.txt
Django==5.0.6
djangorestframework==3.15.1
gunicorn==21.2.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9  
dj-database-url==2.1.0 
requests==2.31.0
yt-dlp==2023.11.16  